update MAC address discovery
This changes how we find MAC addresses, which serve as the middle 48
bits of the ID byte sequence. Previously we used the localhost to
identity a network interface. This has several problems and ultimately
means we would fail to find a MAC more often than we should. To remedy
this, here we instead walk a sequence of all network interfaces
`NetworkInterface/getNetworkInterfaces` yields. From this sequence we
remove any null devices (nil or 0.0.0.0.0.0). If we fail to find a valid
hardware address, we fallback to a random six bytes via `SecureRandom`.