restructure code
This is a fairly large rewrite which aims to simplify the code by
removing the unnecessary `PartialFlake` abstraction. Previously this
was used to build up a Flake progressively as well as expose the last
timestamp seen in an ID for use with the persistent timer. However
such complexity is not needed to achieve the same results and here we
remove much of that work and replace it with a much simpler object
type, `Flake`.

We also have made a few breaking changes, most importantly renaming
`generate` to `generate!` in order to signify the fact this function is
fundamentally mutating global state--this was an oversight previously
and corrected here. In addition to this, `generate!` returns a
ByteBuffer instead of a BigInt. This gives the caller much more
flexibility as to how the ID should be encoded and used but obviously
breaks backwards compatibility. Library consumers should use the helper
function `flake->bigint` if they require the old functionality--simply
wrap calls to `generate!` with `flake->bigint`.

Finally we have also removed the worker ID fallback: the worker ID is
the middle portion of a Flake and is used to identify the host machine
a Flake is generated on. Previously we had a fallback mechanism in place
which would replace a missing hardware address with random bytes.
However it turned out the process for deriving a valid hardware address
was flawed and likely the fallback is often not necessary--additionally
a worker ID can now be provided as a parameter to `generate!` and so a
library-level fallback is unnecessary.