ensure better thread-safety
This patch changes `generate-flake!` so that a potential race condition,
where the stored timestamp is altered by another thread before we swap
on the state atom, is avoided. We do so by moving the logic for checking
timestamps into the function we pass to `swap!` thus ensuring the
operation is thread-safe.

Additionally, `generate!` has been updated to catch
`IllegalStateException` since this is an indication that the current
thread has lost the race to alter the state atom as opposed to a problem
with the system clock. As a follow up to this, there are plans to rework
how we derive timestamps which should minimize risk of clock skew.

If you are generating flakes in a multi-threaded context, i.e. from
multiple threads, this patch is an important bugfix for you.