better time estimation
This patch changes how we determine linear time. Instead of relying on
`System/currentTimeMillis`, which does not necessarily provide a
high-resolution wall clock, we use a custom epoch, derived from
`System/currentTimeMillis` and `System/nanoTime`.

Borrowing from other implementations, e.g. Skuld, we build an epoch
based on the difference between `System/currentTimeMillis` and
`System/nanoTime` where that difference is based on a set of samples--to
guard against clock skew. Timestamps are then built using the mean of
these samples as the epoch.