provide timeout utility
This provides a simple timeout macro which will attempt to execute a
wrapped form into a separate thread. If the timeout, in milliseconds,
elapses before the future can be deref'd, then an exception is raised.
We introduce this to address the fact that `generate!` could, in rare
circumstances, loop in such a manner that the caller would prefer to
timeout the execution rather than wait for its resolution.